Update code organisation. All utils functions and messages are not in separate folder
