+++ telegram-bot/utils/messages.py
@@ -1,0 +1,6 @@
+WELCOME_MESSAGE = 'Hi, welcome to SayNoMore bot. Tell us more about the trip you are planning'
+MORE_INFO_MESSAGE = ('It seems that I did not receive all the necessary information for your request. Please specify '
+                     'the destination city. Additionally, if you have a budget, you can provide that too, although it '
+                     'is not mandatory. Thank you!')
+
+AWAIT_MESSAGE = 'All good, let me prepare a trip plan for you'
